The libified smoketests are now on parity with the seperate ones.
Updated documentation and added better help routines to test_crowbar.sh.

+++ test_framework/test_crowbar.sh
@@ -26,11 +26,12 @@
 
 readonly currdir="$PWD"
 
-# Include common build and smoketest functionality
-if ! [[ $CROWBAR_DIR && -d $CROWBAR_DIR ]]; then
-    echo "Cannot find \$CROWBAR_DIR, we cannot run tests."
-    exit 1
-fi
+# Location of the Crowbar checkout we are building from.
+[[ $CROWBAR_DIR ]] || CROWBAR_DIR="${0%/*}"
+[[ $CROWBAR_DIR = /* ]] || CROWBAR_DIR="$currdir/$CROWBAR_DIR"
+[[ -f $CROWBAR_DIR/build_crowbar.sh && -d $CROWBAR_DIR/.git ]] || \
+    die "$CROWBAR_DIR is not a git checkout of Crowbar!" 
+export CROWBAR_DIR
 
 . "$CROWBAR_DIR/build_lib.sh"
 . "$CROWBAR_DIR/test_lib.sh"
@@ -40,11 +41,57 @@ do_help() {
 $0: Crowbar test framework.
 I smoketest Crowbar to ensure we have certian minimal functionality.
 
-If you run me without options, I will run a smoketest 
-If you run me with run-failed <machine list>, I will try to run
-vms from the last failed run.
+Arguments:
+   cleanup: Will try to clean up from any previous runs that
+            exited abnormally.
+   develop-mode: Causes the framework to pause instead of automatically
+                 exiting whenever a test fails, a timeout is exceeded, or
+                 the test would otherwise exit sucessfully.
+   pause: Causes the framework unconditionally pause after all the tests 
+          have finished.
+   pause-after-admin: Causes the framework to pause after the admin node
+                      has been deployed.
+   admin-only: Only deploy an admin node -- do not spin up compute nodes.
+   manual-deploy: By default, the framework will automatically launch the
+                  crowbar install once the admin node is up using a default
+                  name of admin.smoke.test.  Passing manual-deploy causes
+                  the framework to not do that.
+   use-iso <path-to-iso>: This tells the framework what iso to use to try
+                          to use to deploy Crowbar from.
+   <barclamp name>: Deploy and run any applicable smoketests for <barclamp>.
+   help: Display this help
 
-Read my comments to learn how to tweak me.
+VM Specifications:
+   By default, the smoketest framework will try and deploy a 5 node cluster.
+   
+   The admin node will have 4 CPUs, 4 gigs of RAM, 1 hard drive with
+   15 gigs of disk space, and 2x E1000 nics.
+
+   The compute nodes will have 2 CPUs, 2 gigs of RAM, i hard drive with
+   10 gigs of space for the OS, 2 hard drives with a gig of space, 
+   and 2x E1000 nics.
+
+Networking:
+   The smoketest framework creates 2 bridges, and attaches one nic from each
+   VM to each bridge.  The primary bridge is given an IP address of 
+   192.168.124.1, and the framework assumes that it can contact the admin node
+   at 192.168.124.10.
+
+   The framework will also inject the public key of the user running the
+   smoketest into all of the deployed nodes to enable passwordless SSH access
+   to all the public interfaces on all the deployed nodes.
+
+Logging and Display:
+   By default, the smoketest framework will arrange for all the VMs to log
+   to a serial capture, which is saved in $CROWBAR_DIR/testing.  This is also
+   where error output from the scripts are captured, the VM images and state
+   information is saved, etc.  The framework also runs a screen session to 
+   display various important logs, provide a shell running in the same 
+   environment as the framework, and (if the DISPLAY variable is not present
+   in the environment) to provide access to the consoles of the vms.
+   If DISPLAY is set, the VMs will run using an SDL display.
+
+For more information, please see $CROWBAR_DIR/test_framwwork/README.testing.
 EOF
 }
 
@@ -52,11 +99,19 @@ EOF
 [[ -f $HOME/.ssh/id_rsa.pub ]] || { mkdir -p "$HOME/.ssh"; ssh-keygen -q -f "$HOME/.ssh/id_rsa" -t rsa -N '' ; }
 
 case $1 in
-    cleanup) shift; 
-	for l in "$SMOKETEST_LOCK" "$SMOKETEST_KVM_LOCK" "$SMOKETEST_CLEANUP_LOCK"; do
-	    rm -f "$l"
-	done
-	smoketest_cleanup;;
+    cleanup) shift;
+	    CGROUP_DIR=$(sudo "$SMOKETEST_DIR/make_cgroups.sh" $$ crowbar-test) || \
+		die "Could not mount cgroup filesystem!"
+	    
+	    for smoketest_dir in "$CROWBAR_DIR/testing/"*; do
+		[[ -d $smoketest_dir && $smoketest_dir != */cli ]] || continue
+		smoketest_cleanup
+	    done
+	    for l in "$SMOKETEST_LOCK" "$SMOKETEST_KVM_LOCK" \
+		"$SMOKETEST_CLEANUP_LOCK"; do
+		rm -f "$l"
+	    done;;
     ''|help) do_help;;
+    run-test) shift; run_test "$@" ;; 
     *) run_test "$@" ;;
 esac
